requirements.txt
dart-fss>=0.4.3
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
aiofiles>=23.2.0
python-jose[cryptography]>=3.1.0
